==> R/00_all.qmd <==
## Load libraries

```{r}
library(quarto)
library(here)
library(purrr)
library(stringr)
library(dplyr)

# load functions
source(here("R/99_proj_func.R"))
```

## Rendering the qmd scripts and afterwards moving the created html's to the results folder

```{r}
#|output: false

# List of input files
input_files <- c(
  "R/01_load.qmd",
  "R/02_clean.qmd",
  "R/03_augment.qmd",
  "R/04_describe.qmd",
  "R/05_analysis_1.qmd",
  "R/06_analysis_2.qmd",
  "R/07_analysis_3.qmd"
)

# Creating directories
dir.create("../data")

# Ensure results directory exists
results_dir <- here("results")
#if (!dir.exists(results_dir)) dir.create(results_dir)

# Create a tibble for inputs and outputs
files_tbl <- tibble(
  input = here(input_files),
  output_name = str_replace(basename(input_files), "\\.qmd$", ".html"), # Replace .qmd with .html
  temp_output = file.path(dirname(here(input_files)), output_name),    # Temporary location
  final_output = file.path(results_dir, output_name)                  # Final destination
)

# Render and move files using purrr
files_tbl |> 
  mutate(
    # Render the files
    rendered = map2(input, output_name, ~ quarto::quarto_render(input = .x, output_file = .y)),
    # Move the files to results folder
    moved = map2(temp_output, final_output, ~ file.rename(.x, .y))
  )

# code to copy .png files from results to images
# Define source and destination directories
source_folder <- here("results")  
destination_folder <- here("doc/images")

# Copy .png files from results to images
list.files(source_folder, 
           pattern = "\\.png$", 
           full.names = TRUE) |>   # choose only .png files
  walk(~ file.copy(from = .x, 
                   to = file.path(destination_folder, 
                                  basename(.x)), 
                   overwrite = TRUE))

# Rendering presentation
input_file <- here("doc/presentation.qmd")
output_name <- "presentation.html"
temp_output <- "presentation.html"
final_output <- here("doc/presentation.html")
quarto::quarto_render(input = input_file, output_file = output_name)
file.rename(from = temp_output, to = final_output)


```
