==> R/04_describe.qmd <==
---
title: "04_describe"
format:
  html:
    embed-resources: true
editor: visual
output-dir: "../results/04_describe.html"
---

## Load libraries

```{r}
#| message: false
library("tidyverse")
library("here")
library("dplyr")
```

## Load data

```{r}
describe_dat_aug_01 <- read.csv(here("data/03_dat_aug_01.csv"))
describe_dat_aug_02 <- read.csv(here("data/03_dat_aug_02.csv"))
```

## Data description

### Amount of genes investigated in the experiment

```{r}
# number of genes (observations)
describe_dat_aug_01 |> 
  dim() # return dimensions of df to a vector
```

there is 784 variables which means there is 784 different genes in the experiment

### Amount of classes and genes in each class

```{r}
# identify the classes of genes 
describe_dat_aug_01 |> 
  group_by(code_class) |> # group by gene class
  summarise(n = n()) |>  # calculate number of rows within each gene class group
  print()

# histogram of gene classes and the number of genes in them 
histogram_plot <- describe_dat_aug_01 |> # create histogram
  ggplot(aes(x = code_class, # set gene class on x-axis
             fill = code_class)) + # color by gene class
  geom_bar() + 
  labs(
    title = "Gene distribution of gene classes",
    x = "Classes",
    y = "Number of Genes",
    fill = "Gene Class" 
  ) +
  theme_minimal()

# save plot
ggsave(here("results/04_histogram_plot.png"), 
       plot = histogram_plot, 
       width = 10, 
       height = 6,
       bg = "white")

```

There is four different classes: endogenous, housekeeping, negative and positive.\
There is 750 genes in endogenous, 20 genes in housekeeping, 8 genes in negative and 6 genes in positive

### Amount of pre-treatment and post-treatment experiments

```{r}
describe_dat_aug_01 |> 
    select(starts_with("pre")) |> 
  ncol() |> 
  print()

describe_dat_aug_01 |> 
    select(starts_with("post")) |> 
  ncol() |> 
  print()
```

There are five columns (variables) containing data from pre-treatment experiment and vice verse for post-treatment. This means there was done 10 experiments, five before treatment and five after treatment.

### Density plot of gene expression by code class in pre and post conditions

```{r}
density_gene_class_pre <- 
  ggplot(data = describe_dat_aug_02, # create density plot for pre
         aes(x = mean_pre, 
             fill = code_class)) +
  geom_density(alpha = 0.5) + # a bit transparent
  theme_minimal() +
  scale_x_log10() + # transform x-axis to a logarithmic scale
  scale_y_continuous(breaks = seq(0, 1, by = 0.25)) +
  labs(title = "Density Plot of Pre Gene Class",
       x = "Mean Expression ",
       y = "Density",
       fill = "Gene Class") + 
  theme(plot.title.position = "plot") # make the title go to the very left

density_gene_class_post <- 
  ggplot(data = describe_dat_aug_02,  # create density plot for post
         aes(x = mean_post, 
            fill = code_class)) +
  geom_density(alpha = 0.5) +
  theme_minimal() +
  scale_x_log10() +  # transform x-axis to a logarithmic scale
  scale_y_continuous(breaks = seq(0, 1, by = 0.25)) +
  labs(title = "Density Plot of Post Gene Class",
       x = "Mean Expression ",
       y = "Density",
       fill = "Gene Class") + 
  theme(plot.title.position = "plot") 

# save plots
ggsave(here("results/04_density_plot_pre.png"), 
       plot = density_gene_class_pre, 
       width = 10, 
       height = 6)

ggsave(here("results/04_density_plot_post.png"), 
       plot = density_gene_class_post, 
       width = 10, 
       height = 6)
```

This plot does assess the variation in gene expression across the different code classes and observe any shifts between the "pre" and "post" conditions.

### The most up- and down-regulated genes in terms of fold change before and after treatment of each gene class

```{r}
top2_procent_fold_change <- describe_dat_aug_01 |> 
  group_by(code_class) |> 
 mutate(
  biggest_fold_change = case_when( # add new classifyer
    log2_fold_change <= quantile(log2_fold_change, 
                                 0.02, 
                                 na.rm = TRUE) ~ "Low", # 2% most downregulated
    log2_fold_change >= quantile(log2_fold_change, 
                                 0.98, 
                                 na.rm = TRUE) ~ "High", # 2% most up regulated
    TRUE ~ "middle") # the in betweens are in the catagory "middle"
  ) |> 
  filter(biggest_fold_change %in% c("High", "Low")) |> # only plot of the 2% most down- and up regulated
  ggplot(mapping = aes(x = gene,
                       y = log2_fold_change)) + 
  geom_point(aes(color = code_class, 
                 shape = biggest_fold_change)) + 
  labs(title = "Top 2% Most Upregulated and Downregulated Genes After Treatment by Code Class",
       x = "Genes",
       y = "Fold Change",
       color = "Gene Class",
       shape = "Top or Buttom 2%") + 
  theme(axis.text.x = element_text(angle = 45, hjust = 1),  
        plot.title.position = "plot") 
  
# view plot
print(top2_procent_fold_change)

# save plot
ggsave(here("results/04_top_2_procent_fold_change.png"), 
       plot = top2_procent_fold_change, 
       width = 10, 
       height = 6)
```

### Which genes are most up- and down regulated ?

```{r}
describe_dat_aug_01 |> 
  group_by(code_class) |> # group by gene class
 mutate( # Classify the 'log2_fold_change' values into categories 
  biggest_fold_change = case_when(
    log2_fold_change <= quantile(log2_fold_change,  # the bottom 2% are labeled as "Low"
                                 0.02, na.rm = TRUE) ~ "Low",
    log2_fold_change >= quantile(log2_fold_change, # the top 2% are labeled as "High"
                                 0.98, na.rm = TRUE) ~ "High",
    TRUE ~ "middle") # the rest are labeled as "High"
  ) |> 
  filter(biggest_fold_change %in% c("High", "Low")) |> # keep only high and low values
  arrange(desc(log2_fold_change)) |> # arrange from highest log2 fold change to lowest 
  print()

```
