==> R/06_analysis_2.qmd <==
---
title: "06_analysis_2"
format:
  html:
    embed-resources: true
editor: visual
output-dir: "../results/06_analysis_2.html"
---

## Load Libraries

```{r}
#| message: false
library("tidyverse")
library("here")
library("dplyr")
library("ggplot2")
```

## Load data

```{r}
log2_fold_change <- read.csv(here("data/03_dat_aug_02.csv"))
```

## Analysis

The analysis represents the distribution of log2 fold changes for a specific gene class

```{r}
boxplot_of_gene_class_log2fc <- log2_fold_change |>
  ggplot(aes(
    x = code_class, # set code_class as x-axis
    y = log2_fold_change,  # set log2_fold_change as y-axis
    fill = code_class)) + # color by code_class
  labs(
    title = "Log2 Fold Change Differences Between Gene Classes",
    x = NULL,
    y = "Log2 Fold Change",
    fill = "Gene Class"
    ) +
  scale_fill_brewer(palette = "Paired") +
  geom_boxplot() +
  scale_x_discrete(expand = expansion(mult = c(0.2, 0.2))) +
  theme(
    axis.text.x = element_text(size = 8,
                               angle = 0,
                               vjust = 1,
                               hjust = 0.5),
    axis.text.y = element_text(size = 8), # rotate labels if needed
    axis.title.x = element_text(margin = margin(t = 10)),
    axis.title.y = element_text(size = 8),
    plot.title.position = "plot"
  )

# save plot
ggsave(here("results/06_boxplot_of_gene_class_log2fc.png"), plot = boxplot_of_gene_class_log2fc, width = 10, height = 6)
```
